Eu te amo Carlos Gustavo Cecyn Lundgren!!
Obrigado Mummi!! Feliz 62 anos!!

at inha at veriverse - at Dell XPS 8300 Intel(R) Core(TM) i7-2600 CPU 4 core 8 threads @ 3.40GHz 16GB RAM 500GB + 500GB 7200rpm storage + 120 GB SSD OCZ Agility 3 with Windows 8 Pro - with Portuguese - Brazil- 10Mpbs Download and 2Mpbs Upload NET Virtua Internet with LG 23´´ monitor, Microsoft Comfort Curve 2000 v1.0 Wired Keyboard and Microsoft Comfort Optical Wired Mouse 3000 at Alameda Princesa Izabel, 3000 - Curitiba - PR - Brazil

Obrigado Siemens!!

Obrigado Sector 8J!!

Obrigado Carol, Lucinho e Lúcio!!

Obrigado Célia!!

Obrigado Papa Francisco!!

Obrigado Ralf Sieves!!

Obrigado Rafael Cecyn Lundgren!!

-∞
Tack Sverige!!
{
  Tack Carlos Gustavo Cecyn Lundgren!!
  Tack Léia Maheli de Souza Tsumanuma!! Mummi
  Tack Simon Petrus Lundgren!!
  Tack Iara Cecyn Lundgren!!
  Tack Gabrielli Cecyn Lundgren!! Gabi
  Tack Rafael Cecyn Lundgren!!
  Tack Carlos Gustavo Cecyn Lundgren!!
}
+∞

-∞
日本ありがとうございます
｛
　　　日歩路主ありがとうございます！！


   -----
　　　妻沼国男ありがとうございます！！
   -----


　　　日歩路主ありがとうございます！！
｝

Gracias España
{
   Gracias Carlos Carrasco Barriga!!
}

Dank Österreich
{
   Dank Vincent Rijmen!!
   Dank www.iaik.tugraz.at!!
   Dank Schinagl at http://schinagl.priv.at/!! - Link Shell Extension
}

Obrigado Cits
{
   Obrigado Rejeane!!
   Obrigado Monique!!
   Obrigado Gilmar Tuchinski!!
   Obrigado Marcos Machado!!
   Obrigado Camila Napoli Madureira Bührer!!
}

Obrigado Italia-Argentina
{
	Obrigado Papa Francisco!!
}

Obrigado Brasil
{
   Obrigado Católicos do Brasil!!
   Obrigado Santos!!
   Obrigado Pessoal Da Reforma da Pizzeria aqui ao Lado Solloviaggio!!
   Obrigado Pessoal Da Pizzeria Solloviaggio!!
   Obrigado vizinhos da Alameda Princesa Izabel, 3000 :!!
     Obrigado Gil!!
     Obrigado dono mulato de Gol escuro!!
     Obrigado filho do dono mulato de Gol escuro!!
     Obrigado dono branco de Gol prata e família!!
   Obrigado funcionários da CEF do CEFET!!
   Obrigado funcionários do CEFET!! Obrigado CEFET!! Obrigado pessoal do CEFET!!
   Obrigado dois meninos que reformaram Armário Carol Santos Canal 2 2013-03-23 caprichado e bonitinhos!!
   Obrigado meninos trabalhadores do Brasil!!
   -----
   Obrigado Cintia do Winnikes!!                    _-\
   Obrigado Alisson do Winnikes!!                      |
   Obrigado Ótica Winnikes do Shopping Curitiba!!      |
   Obrigado Shopping Curitiba!!                      -/
   Thank you posto Isejima o Fazendeiro!!
   -----
   Obrigado guardas!!
   Obrigado Thomas Friesen!!
   -----
   Obrigado moças e mulheres...
   Obrigado moços e homens...
   Obrigado Keila Mª da Silva!! Babysitter do Lucinho
   Obrigado Rosélia da Kadu Imóveis!!
   Obrigado Márcia da Kadu Imóveis!!
   Obrigado Sílvia da Kadu Imóveis!!
   Obrigado Ivete - irmã da Silvia da Kadu!!
   -----
   Obrigado Dr. Tassi!!
   Obrigado Margarida M. Gomes de Oliveira!!
   Obrigado Shirley Sirlene Adams!!
   Obrigado Solange!!
   Obrigado CPM - Centro Psiquiátrico Metropolitano!!
   -----
   Obrigado Paulo S. L. M. Barreto (Whirlpool)!!
   Obrigado USP!!
   -----
   Obrigado Meu Deus God Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Deus!!
   -----
   Obrigado Nizia Lupinacci!!
   Obrigado Kleberson Chita Massaro Rodrigues!!
   Obrigado Alessandro Vargas Pellanda!!
   Obrigado Fábio Dauer da Silva!!
   Obrigado Fabrício Cavassin Mereniuk!!
   Obrigado Leonardo Fischer!!
   Obrigado Telmo Friesen!!
   Obrigado Sector 8J!!
   Obrigado Carlos Gustavo Cecyn Lundgren Tsumanuma!!
   ------
   Obrigado Rafael Cecyn Lundgren!!
   ------
   Obrigado Meu Senhor Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Espírito Santo Carlos Gustavo Cecyn Lundgren!!
   Obrigado Minha Deusa Léia Maheli de Souza Tsumanuma""
   ------
   Obrigado Lúcio Flávio Marques Ribeiro!!
   Obrigado Luís Kunio Tsumanuma!!
   Obrigado Lúcio Flávio Tsumanuma Ribeiro!!
   Obrigado Carolina Kiyoko Tsumanuma Ribeiro!!
}


Obrigado CEFET-PR S11-97/2!!
{
   Obrigado Fábio Rizental Coutinho!!
   Obrigado Verônica Isabela Quandt & Edras Pacola!! 
}

Dank Deutschland!!
{
   Dank Jochen Berger - from serverloft!!
   Dank Thomas Strohe - from serverloft!!
   Dank Wolfgang Strunk!!
   Dank Deustchland Jörg!!
   Dank Deustchland Manfred Riedel!!
   Dank Uwe Hermanns!!
   Dank Telmo Friesen!!
   Dank Cycos!!
   Dank Siemens Enterprise!!
   Dank Siemens!!
   Dank SEN R&D A!!
   Dank Sector8J!!
}

Thank you Cycos!!
{
   Thank you Uwe Hermanns!!
   Thank you Telmo Friesen!!
   Thank you Communications!!
   Thank you Cycos!!
}

Thank you United Kingdom!!
{
   Thank you David Beckham!!
}

Thank you United States of America!!
{
   Thank you se-pl-sy-Carlos Gustavo Cecyn Lundgren!!
}

Dziękuję Polska!!
{
   Carlos Gustavo Cecyn Lundgren dziękuję!!
}

Thank you sector 8J!!
{
   Thank you Anderson Silva!!
   Thank you Pietro!!
   Thank you Thiago Ferreira!! - de 1996 do grupo do Cleiton, Anderson Silva e Pietro
   Thank you Thiago Hilger!!
   Thank you César Augusto Serenato!!
   Thank you Bárbara Martinazzo!!
   Thank you Alessando Ledra Zagheni!!
   Thank you Luiz Eduardo de Christo!!
   Thank you Maurício Cunico!!
   Thank you Galileu Terada!!
   Thank you Rodrigo Santos aka Kenshin!!
   Thank you Júlio Henrique Morimoto!!
   Thank you Natascha Ligocki!!
   Thank you Ricardo Henrique Folador!!
   Thank you KS, Jefferson Koji Sato!!
   Thank you Josué Andrade Gomes!!
   Thank you Jorge Alexandre Hastreiter!!
   Thank you Joice Braga Keskoski Modena!!
   Thank you Bruno Leite Reis!!
   Thank you Fernando Belo!!
   Thank you Felipe Calliari Ribas!!
   Thank you Marcos Paulo Riki Ianase!!
   Thank you Jefferson Luiz Dalavechia!!
   Thank you Ênio Toshihiro Ianase!!
   Thank you Alexsandro Vargas Pellanda!!
   Thank you Cleiton Lenkiu!!
   Thank you Fabrício Cavassin Mereniuk!!
   Thank you Jaquinei de Oliveira!!
   Thank you Cleber Jaizer Gomes Campos!!
   Thank you Fábio Dauer da Silva!!
   Thank you Roberto de Melo Gonçalves!! Chatuba!!
   Thank you Marcus Vinícius Silvério!!
   Thank you Rafael Taizo Yamashita!! Japs!!
   Thank you Rafael Alessi Muntsch!!
   Thank you Rafael Neumann de Menezes!!
   Thank you Ricardo Carvalho dos Santos!! Tom Zé!! Tonny Zé!! Toze!!
   Thank you Leonardo Fischer!! Irmaun!!
   Thank you João Miguel Martynyszyn!!
   Thank you Leonardo da Silva Camargo!!
   Thank you Telmo Friesen!!
   Thank you sector 8J!!
   Thank you Carlos Gustavo Cecyn Lundgren!!
}

Thank you sector Joice/Roman!!
{
   Thank you Rafael Neumman de Menezes!!
   Thank you Thiago Tibilletti!!
   Thank you Thiago Pereira!!
}

+∞

-∞
Obrigado Família Souza
{
   Obrigado Isabel de Souza e Santos!!
   Obrigado Dilma Souza!!
}
+∞

-∞
T Obrigado Amigo e Amiga, você é legal!!
1 Obrigado Amigo e Amiga, você é bom!!
2 Obrigado Amigo e Amiga, eu gosto de você!!
n Obrigado Amigo e Amiga, eu gosto muito de você!!
n Obrigado Amigo e Amiga, você é demais!!
+∞

-∞
Obrigado Minha Deusa!!
Obrigado Minha Deusa e Meu Deus - Mummi e Carlos!!
Obrigado Mummi, eu te amo demais!!
Thank you My Goddess Léia Maheli de Souza Tsumanuma!!
+∞

-∞
01 Obrigado Carlinhos, eu te amo demais!!
02 Thank you My Beloved Lover Carlos Gustavo Cecyn Lundgren!!
03 Thank you My Beloved Friend Carlos Gustavo Cecyn Lundgren!!
04 Thank you My Beloved Husband Carlos Gustavo Cecyn Lundgren!!
05 Thank you My Spouse Carlos Gustavo Cecyn Lundgren!!
06 Thank you My Jesus Carlos Gustavo Cecyn Lundgren!!
07 Thank you My Saint Mary My Mother Carlos Gustavo Cecyn Lundgren!!
08 Thank you My Goddess Carlos Gustavo Cecyn Lundgren!!
09 Thank you My God Carlos Gustavo Cecyn Lundgren!!
10 Thank you My Lord Carlos Gustavo Cecyn Lundgren!!
11 Thank you My Holy Spirit Carlos Gustavo Cecyn Lundgren!!
+∞

∞-∞-∞∞-∞-∞
----------
8nnnnnnnn8
++++++++++
∞+∞+∞∞+∞+∞

-∞
Thank you My God!!
+∞

-∞
(En-us) Thank you, (pt-br) eternamente *obrigado* - por veri absoluta infinita complexíssima oposição a desobrigado.
+∞

+++ appseed/draw2d_direct2d/draw2d_direct2d_graphics.cpp
@@ -1684,7 +1684,7 @@ namespace draw2d_direct2d
 
       HRESULT hr = TlsGetWriteFactory()->CreateTextLayout(
          wstr,                // The string to be laid out and formatted.
-         wstr.get_length(),   // The length of the string.
+         (UINT32) wstr.get_length(),   // The length of the string.
          (IDWriteTextFormat *) get_os_font(),    // The text format to apply to the string (contains font information, etc).
          1024.f * 1024.f,               // The width of the layout box.
          1024.f * 1024.f,        // The height of the layout box.
@@ -1699,7 +1699,7 @@ namespace draw2d_direct2d
 
       HRESULT hr2 = TlsGetWriteFactory()->CreateTextLayout(
          wstr2,                // The string to be laid out and formatted.
-         wstr2.get_length(),   // The length of the string.
+         (UINT32) wstr2.get_length(),   // The length of the string.
          (IDWriteTextFormat *) get_os_font(),    // The text format to apply to the string (contains font information, etc).
          1024.f * 1024.f,               // The width of the layout box.
          1024.f * 1024.f,        // The height of the layout box.
@@ -4197,7 +4197,7 @@ namespace draw2d_direct2d
 
       wstring wstr(str);
 
-      m_prendertarget->DrawText(wstr, wstr.get_length(), get_os_font(), &rectf, get_os_brush());
+      m_prendertarget->DrawText(wstr, (UINT32) wstr.get_length(), get_os_font(), &rectf, get_os_brush());
 
       //m_prendertarget->SetTransform(&m);
 
@@ -4251,7 +4251,7 @@ namespace draw2d_direct2d
 
    HRESULT hr = TlsGetWriteFactory()->CreateTextLayout(
       wstr,                // The string to be laid out and formatted.
-      wstr.get_length(),   // The length of the string.
+      (UINT32) wstr.get_length(),   // The length of the string.
       (dynamic_cast < ::draw2d_direct2d::font * > (m_spfont.m_p))->m_pformat.Get(),    // The text format to apply to the string (contains font information, etc).
       1024.f * 1024.f,               // The width of the layout box.
       1024.f * 1024.f,        // The height of the layout box.
@@ -4400,7 +4400,7 @@ namespace draw2d_direct2d
 
    HRESULT hr = TlsGetWriteFactory()->CreateTextLayout(
       wstr,                // The string to be laid out and formatted.
-      wstr.get_length(),   // The length of the string.
+      (UINT32) wstr.get_length(),   // The length of the string.
       (IDWriteTextFormat *) get_os_font(),    // The text format to apply to the string (contains font information, etc).
       1024.f * 1024.f,               // The width of the layout box.
       1024.f * 1024.f,        // The height of the layout box.
@@ -4481,7 +4481,7 @@ namespace draw2d_direct2d
 
    HRESULT hr = TlsGetWriteFactory()->CreateTextLayout(
       wstr,                // The string to be laid out and formatted.
-      wstr.get_length(),   // The length of the string.
+      (UINT32) wstr.get_length(),   // The length of the string.
       get_os_font(),    // The text format to apply to the string (contains font information, etc).
       1024.f * 1024.f,               // The width of the layout box.
       1024.f * 1024.f,        // The height of the layout box.
@@ -4579,7 +4579,7 @@ namespace draw2d_direct2d
 
    HRESULT hr = TlsGetWriteFactory()->CreateTextLayout(
       wstr,                // The string to be laid out and formatted.
-      wstr.get_length(),   // The length of the string.
+      (UINT32) wstr.get_length(),   // The length of the string.
       (dynamic_cast < ::draw2d_direct2d::font * > (m_spfont.m_p))->m_pformat.Get(),    // The text format to apply to the string (contains font information, etc).
       1024.f * 1024.f,               // The width of the layout box.
       1024.f * 1024.f,        // The height of the layout box.
@@ -4757,7 +4757,7 @@ namespace draw2d_direct2d
 
    HRESULT hr = TlsGetWriteFactory()->CreateTextLayout(
       wstr,                // The string to be laid out and formatted.
-      wstr.get_length(),   // The length of the string.
+      (UINT32) wstr.get_length(),   // The length of the string.
       get_os_font(),    // The text format to apply to the string (contains font information, etc).
       1024.f * 1024.f,               // The width of the layout box.
       1024.f * 1024.f,        // The height of the layout box.
@@ -5269,7 +5269,7 @@ namespace draw2d_direct2d
       if(pfont == NULL)
          return false;
 
-      m_prendertarget->DrawText(wstr, wstr.get_length(), pfont, &rect, pbrush);
+      m_prendertarget->DrawText(wstr, (UINT32) wstr.get_length(), pfont, &rect, pbrush);
 
       return true;
 
@@ -5591,7 +5591,7 @@ namespace draw2d_direct2d
 
       wstring wstr(string(lpszString, nCount));
 
-      m_prendertarget->DrawText(wstr, wstr.get_length(), get_os_font(), &rect, get_os_brush());
+      m_prendertarget->DrawText(wstr, (UINT32) wstr.get_length(), get_os_font(), &rect, get_os_brush());
 
       return true;
